requirements.txt
openai==1.35.13
Pillow==9.2.0
python-dotenv==1.0.1
streamlit==1.36.0
